fix(android): compilation errors
Fixes an error in `FlutterBoostActivity` and `FlutterBoostFragment` in which a call to `FlutterEngine().getActivityControlSurface().attachToActivity` was passing an `Activity` object as the first parameter, instead of the expected `ExclusiveAppComponent<Activity>`.

Fixes an error in `LifecycleView` in which the class was missing the implementation of a method defined in one of its super-types.